% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\name{add}
\alias{add}
\title{A simple adding functions}
\usage{
add(x, y)
}
\arguments{
\item{x}{a number}

\item{y}{another number}
}
\description{
An example adding function
}
\examples{
add(1, 2)
}

